Moves styles and colors to a constants.ts file
