--- src/components/Colors.tsx
@@ -1,9 +1,0 @@
-export const Colors = {
-    white: 'white',
-    black: 'black',
-    green: '#50E3C2',
-    maroon: '#800000',
-    darkblue: '#000080',
-    needText: '#505E65',
-    separatorColor: 'rgba(0,0,0,0.1)'
-}
